confirmation of cancelling vote
git-svn-id: http://opensvn.csie.org/pttbbs/trunk/pttbbs@3381 63ad8ddf-47c3-0310-b6dd-a9e9d9715204

+++ mbbsd/vote.c
@@ -1,4 +1,4 @@
-/* $Id: vote.c 3341 2006-04-08 14:58:06Z kcwu $ */
+/* $Id: vote.c 3381 2006-07-22 04:46:30Z wens $ */
 #include "bbs.h"
 
 #define MAX_VOTE_NR	20
@@ -541,6 +541,11 @@ vote_view(vote_buffer_t *vbuf, const char *bname, int vote_index)
     getdata(b_lines - 1, 0, "(A)取消投票 (B)提早開票 (C)繼續？[C] ", genbuf,
 	    4, LCECHO);
     if (genbuf[0] == 'a') {
+	getdata(b_lines - 1, 0, "請再次確認取消投票 (Y/N) [N] ", genbuf,
+		4, LCECHO);
+	if (genbuf[0] == 'n')
+	    return FULLUPDATE;
+
 	setbfile(buf, bname, vbuf->control);
 	unlink(buf);
 	setbfile(buf, bname, vbuf->flags);
